requirements.txt
# Dataset assembly — use with: .venv\Scripts\activate then pip install -r requirements.txt
pypdf>=3.0.0
pandas>=1.5.0
numpy>=1.21.0
networkx>=2.8
requests>=2.28.0
pyyaml>=6.0
tqdm>=4.64.0
scikit-learn>=1.2.0
biopython>=1.80
lxml>=4.9.0
